rename type `Renaming` into `Arc`
An arc is defined a a couple (source path, target path), and can safely be fed to an actual renaming operation.
The new term (from graph theory) is far from ideal, but it is short. The goal is to disambiguate "renaming", which until now was used both as a type name (data) and an operation.